docker/hw2/stocks_products/requirements.txt
Django==3.2.9
django_filter==21.1
djangorestframework~=3.13.1
psycopg2-binary==2.9.2